do some change in views

+++ BACKEND/E-SHOP/myapp/models.py
@@ -31,7 +31,7 @@ class Cart(models.Model):
     
 class Order(models.Model):
     user = models.ForeignKey(User,on_delete=models.CASCADE)
-    total_price = models.DecimalField(max_digits=10,decimal_places=2)
+    total_price = models.FloatField()
     is_paid = models.BooleanField(default=False)
     created_at=models.DateTimeField(auto_now=True)
 
@@ -42,7 +42,7 @@ class OrderItem(models.Model):
     order = models.ForeignKey(Order,on_delete=models.CASCADE)
     product = models.ForeignKey(Product,on_delete=models.CASCADE)
     quantity = models.PositiveIntegerField(default=1)
-    price = models.IntegerField()
+    price = models.FloatField()
